backend_for_reference/src/backend/routing.py
# mysite/routing.py
from channels.auth import AuthMiddlewareStack
from channels.routing import ProtocolTypeRouter, URLRouter
from django.urls import re_path
#from cons import consumers 

websocket_urlpatterns = [
    re_path(r'ws/path/(?P<var>\w+)/$', consumers.basicConsumer),
]

application = ProtocolTypeRouter({
    'websocket': AuthMiddlewareStack(
        URLRouter(
            websocket_urlpatterns
        )
    ),
})